+++ zcash_primitives/CHANGELOG.md
@@ -7,9 +7,20 @@ and this library adheres to Rust's notion of
 
 ## [Unreleased]
 ### Added
-- `zcash_primitives::sapling::keys::DiversifiableFullViewingKey`
-- `zcash_primitives::sapling::keys::Scope`
+- `zcash_primitives::legacy::AccountPrivKey::{to_bytes, from_bytes}`
 - `zcash_primitives::sapling::NullifierDerivingKey`
+- Added in `zcash_primitives::sapling::keys`
+  - `DecodingError`
+  - `DiversifiableFullViewingKey`
+  - `Scope`
+  - `ExpandedSpendingKey::from_bytes`
+  - `ExtendedSpendingKey::{from_bytes, to_bytes}`
+- Added in `zcash_primitives::zip32`
+  - `ChainCode::as_bytes`
+  - `DiversifierKey::{from_bytes, as_bytes}`
+  - `DiversifierIndex::{as_bytes}`
+  - `ExtendedSpendingKey::{from_bytes, to_bytes}`
+  - Implementations of `From<u32>` and `From<u64>` for `DiversifierIndex`
 
 ### Changed
 - `zcash_primitives::sapling::ViewingKey` now stores `nk` as a
@@ -17,6 +28,8 @@ and this library adheres to Rust's notion of
 - The signature of `zcash_primitives::sapling::Note::nf` has changed to
   take just a `NullifierDerivingKey` (the only capability it actually required)
   rather than the full `ViewingKey` as its first argument.
+- Made the internals of `zip32::DiversifierKey` private; use `from_bytes` and
+  `as_bytes` on this type instead.
 
 ## [0.7.0] - 2022-06-24
 ### Changed